+++ src/App.js
@@ -1,5 +1,4 @@
 import { Route, Routes } from 'react-router-dom';
-import Main from '@pages/Main';
 import ThreeDayDeck from './pages/ThreeDayDeck';
 import ThreeDayCard from '@pages/ThreeDayCard';
 import PermanentCard from '@pages/PermanentCard';
@@ -9,11 +8,15 @@ import QnA from '@pages/QnA';
 import Login from '@pages/Login';
 import KakaoRedirectPage from '@pages/KakaoRedirectPage';
 import SignUp from '@pages/SignUp';
+import SelectProject from '@components/main/SelectProject';
+import StartPage from '@pages/StartPage';
+import Main from '@pages/Main';
 
 function App() {
   return (
     <Routes>
-      <Route path='/' element={<Main />} />
+      <Route path='/' element={<StartPage />} />
+      <Route path='/main' element={<Main />} />
       <Route path='/3day' element={<ThreeDayDeck />} />
       <Route path='/3day/card' element={<ThreeDayCard />} />
       <Route path='/permanent' element={<PermanentDeck />} />
@@ -23,6 +26,7 @@ function App() {
       <Route path='/login' element={<Login />} />
       <Route path='/oauth/kakao/callback' element={<KakaoRedirectPage />} />
       <Route path='/signup' element={<SignUp />} />
+      <Route path='/SelectProject' element={<SelectProject />} />
     </Routes>
   );
 }
